requirements.txt
flask-script
flask-bootstrap
flask-wtf
flask-login
flask-sqlalchemy
simplejson
pymysql